clang-tidy

+++ src/extension/IOController_Posix.hpp
@@ -149,7 +149,7 @@ namespace extension {
                             for (auto it = range.first; it != range.second; ++it) {
 
                                 // Load in the relevant events that happened into the waiting events
-                                it->waiting_events = it->waiting_events | (it->events & fd.revents);
+                                it->waiting_events = short(it->waiting_events | (it->events & fd.revents));
                             }
                         }
                     }
